docs/slides/Lec21/vetiver/docker/vetiver_requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.12
# by the following command:
#
#    pip-compile --output-file=docker/vetiver_requirements.txt /var/folders/v7/wrxd7cdj6l5gzr0191__m9lr0000gr/T/tmpo9b08hgw.in
#
annotated-types==0.7.0
    # via pydantic
anyio==4.9.0
    # via
    #   httpx
    #   starlette
appdirs==1.4.4
    # via pins
build==1.2.2.post1
    # via pip-tools
certifi==2025.1.31
    # via
    #   httpcore
    #   httpx
    #   requests
charset-normalizer==3.4.1
    # via requests
click==8.1.8
    # via
    #   pip-tools
    #   rsconnect-python
    #   uvicorn
fastapi==0.115.12
    # via vetiver
fsspec==2025.3.1
    # via pins
h11==0.14.0
    # via
    #   httpcore
    #   uvicorn
httpcore==1.0.7
    # via httpx
httpx==0.28.1
    # via vetiver
humanize==4.12.2
    # via pins
idna==3.10
    # via
    #   anyio
    #   httpx
    #   requests
importlib-metadata==8.6.1
    # via pins
importlib-resources==6.5.2
    # via pins
jinja2==3.1.6
    # via pins
joblib==1.4.2
    # via
    #   pins
    #   scikit-learn
    #   vetiver
markupsafe==3.0.2
    # via jinja2
narwhals==1.32.0
    # via plotly
nest-asyncio==1.6.0
    # via vetiver
numpy==2.2.4
    # via
    #   pandas
    #   scikit-learn
    #   scipy
    #   vetiver
packaging==24.2
    # via
    #   build
    #   plotly
pandas==2.2.3
    # via
    #   pins
    #   vetiver
pins==0.8.7
    # via vetiver
pip-tools==7.4.1
    # via vetiver
plotly==6.0.1
    # via vetiver
pydantic==2.11.1
    # via
    #   fastapi
    #   vetiver
pydantic-core==2.33.0
    # via pydantic
pyjwt==2.10.1
    # via rsconnect-python
pyproject-hooks==1.2.0
    # via
    #   build
    #   pip-tools
python-dateutil==2.9.0.post0
    # via pandas
python-dotenv==1.1.0
    # via vetiver
pytz==2025.2
    # via pandas
pyyaml==6.0.2
    # via pins
requests==2.32.3
    # via
    #   pins
    #   vetiver
rsconnect-python==1.25.2
    # via vetiver
scikit-learn==1.6.1
    # via
    #   -r /var/folders/v7/wrxd7cdj6l5gzr0191__m9lr0000gr/T/tmpo9b08hgw.in
    #   vetiver
scipy==1.15.2
    # via scikit-learn
semver==3.0.4
    # via rsconnect-python
six==1.17.0
    # via python-dateutil
sniffio==1.3.1
    # via anyio
starlette==0.46.1
    # via fastapi
threadpoolctl==3.6.0
    # via scikit-learn
typing-extensions==4.13.0
    # via
    #   anyio
    #   fastapi
    #   pydantic
    #   pydantic-core
    #   rsconnect-python
    #   typing-inspection
typing-inspection==0.4.0
    # via pydantic
tzdata==2025.2
    # via pandas
urllib3==2.3.0
    # via requests
uvicorn==0.34.0
    # via vetiver
vetiver==0.2.6
    # via -r /var/folders/v7/wrxd7cdj6l5gzr0191__m9lr0000gr/T/tmpo9b08hgw.in
wheel==0.45.1
    # via pip-tools
xxhash==3.5.0
    # via pins
zipp==3.21.0
    # via importlib-metadata

# The following packages are considered to be unsafe in a requirements file:
# pip
# setuptools
